requirements.txt
chainlit==1.3.2
openai==1.54.4
python-dotenv==1.0.1

-e .